New field option, namingPattern
