+++ controllers/dataController.js
@@ -10,6 +10,7 @@ const catchAsync = require('./../utils/catchAsync');
 const AppError = require('./../utils/appError');
 const APIFeatures = require('./../utils/apiFeatures');
 const { replaceAllDataIds } = require('./eventLogController');
+const { getPopulateObj } = require('../utils/misc');
 
 //if collectionName is set, then save that Model as a res.locals.Model
 exports.setModel = (req, res, next) => {
@@ -223,39 +224,6 @@ const getDataSummarySchema = (collectionName, projectName, type) => {
   return null;
 };
 
-// popArr =["experiments_id","experiments_id.projects_id", "experiments_id.test_id"]
-// returns popObj:{
-//   path: 'experiments_id',
-//   populate: { path: 'projects_id test_id' }
-// }
-
-exports.getPopulateObj = popArr => {
-  let popObj = {};
-  // * prepare popObj
-  for (let i = 0; i < popArr.length; i++) {
-    if (!popArr[i].match(/\./)) {
-      // parent fields without dots
-      if (popObj.path) popObj.path += ` ${popArr[i]}`;
-      if (!popObj.path) popObj.path = popArr[i];
-    } else {
-      const fields = popArr[i].split('.');
-      const level = fields.length;
-      const lastfield = fields[level - 1];
-      const populates = `populate${'.populate'.repeat(level - 2)}`; // nested populates
-      // lodash used for setting multiple levels of object with dot notation
-      if (!_.get(popObj, `${populates}.path`)) {
-        // set `path` value with `lastfield`
-        _.set(popObj, `${populates}.path`, lastfield);
-      } else {
-        // get last `path` value and concat with `lastfield`
-        const path = _.get(popObj, `${populates}`).path;
-        _.set(popObj, `${populates}.path`, `${path} ${lastfield}`);
-      }
-    }
-  }
-  return popObj;
-};
-
 const parseSummarySchema = async (collectionName, projectName, format, type) => {
   // e.g. const schema = {
   //    collection: 'sample',
@@ -301,7 +269,7 @@ const parseSummarySchema = async (collectionName, projectName, format, type) => 
   select = uniqueNames.join(' ');
   const popArr = schema.populate.replace(/\s+/g, ' ').split(' ');
   // * prepare popObj
-  const popObj = exports.getPopulateObj(popArr);
+  const popObj = getPopulateObj(popArr);
 
   // * prepare `rename` function
   // e.g. project obj: {
@@ -333,9 +301,6 @@ exports.getDataSummaryDoc = async (type, req, res, next) => {
       req.params.format,
       type
     );
-    console.log('targetCollection', targetCollection);
-    console.log('popObj', popObj);
-    console.log('modelObj', modelObj);
     if (!modelObj[targetCollection]) return null;
     const query = modelObj[targetCollection].find({});
     if (res.locals.Perms) {
